+++ gehomesdk/erd/converters/hood/erd_hood_light_level_new_converter.py
@@ -1,0 +1,13 @@
+from ..abstract import ErdReadWriteConverter
+from ..primitives import *
+from ...values.hood import *
+
+class ErdHoodLightLevelNewConverter(ErdReadWriteConverter[ErdHoodLightLevelNew]):
+    def erd_decode(self, value: str) -> ErdHoodLightLevelNew:
+        try:
+            return ErdHoodLightLevelNew(erd_decode_int(value))
+        except:
+            return ErdHoodLightLevelNew.OFF
+
+    def erd_encode(self, value: ErdHoodLightLevelNew) -> str:
+        return erd_encode_int(value.value, 1)