+++ gehomesdk/erd/erd_configuration.py
@@ -341,8 +341,8 @@ _configuration = [
     ErdConfigurationEntry(ErdCode.HOOD_ACTUAL_FAN_SPEED, ErdHoodFanSpeedConverter(), ErdCodeClass.FAN),
     ErdConfigurationEntry(ErdCode.HOOD_REQUESTED_FAN_SPEED, ErdHoodFanSpeedConverter(), ErdCodeClass.FAN),
     ErdConfigurationEntry(ErdCode.HOOD_AVAILABLE_FAN_SPEEDS, ErdIntConverter(length=1), ErdCodeClass.HOOD_SENSOR, ErdDataType.INT),
-    ErdConfigurationEntry(ErdCode.HOOD_ACTUAL_LIGHT_LEVEL, ErdHoodLightLevelConverter(), ErdCodeClass.LIGHT),
-    ErdConfigurationEntry(ErdCode.HOOD_REQUESTED_LIGHT_LEVEL, ErdHoodLightLevelConverter(), ErdCodeClass.LIGHT),
+    ErdConfigurationEntry(ErdCode.HOOD_ACTUAL_LIGHT_LEVEL, ErdHoodLightLevelNewConverter(), ErdCodeClass.LIGHT),
+    ErdConfigurationEntry(ErdCode.HOOD_REQUESTED_LIGHT_LEVEL, ErdHoodLightLevelNewConverter(), ErdCodeClass.LIGHT),
     ErdConfigurationEntry(ErdCode.HOOD_AVAILABLE_LIGHT_LEVELS, ErdIntConverter(length=1), ErdCodeClass.HOOD_SENSOR, ErdDataType.INT),
 
     #Opal Ice Maker